Login verification refactored, moved interfaces
