+++ src/app/models/user.interface.ts
@@ -1,0 +1,6 @@
+export interface User {
+  id: number;
+  name: string;
+  email: string;
+  password: string;
+}
