+++ src/app/models/verification-result.interface.ts
@@ -1,0 +1,5 @@
+export interface VerificationResult {
+  success: boolean;
+  message: string;
+  userId?: number;
+}
