+++ src/app/services/data.service.ts
@@ -1,9 +1,10 @@
 import { inject, Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Observable, of } from 'rxjs';
-import { catchError, map, tap } from 'rxjs/operators';
-import { User } from './in-memory-data.service';
+import { catchError, filter, map, tap } from 'rxjs/operators';
 import { Router } from '@angular/router';
+import { User } from '../models/user.interface';
+import { VerificationResult } from '../models/verification-result.interface';
 
 @Injectable({
   providedIn: 'root',
@@ -20,18 +21,25 @@ export class DataService {
     return this.http.get<User[]>(this.apiUrl).pipe(catchError(this.handleError('getUsers', [])));
   }
 
-  checkData(email: string | null | undefined, password: string | null | undefined, Users: User[]) {
-    const found = Users.find((e) => e.email === email);
-    if (found) {
-      if (found.password === password) {
-        this.router.navigate(['/main']);
-        return {success: true, message: `Hello ${found.name}`, userId: found.id};
-      } else {
-        return {success: false, message: "Password is wrong!"};
-      }
-    }
-    console.log(`Введён email ${email}, найден еmail ${found}`)
-    return {success: false, message: 'Wrong email'};
+  checkData(email?: string | null, password?: string | null): Observable<VerificationResult> {
+    return this.getUsers().pipe(
+      filter((users) => !!users?.length),
+      map((users) => {
+        const userForCheck = users.find((user) => user.email === email);
+
+        if (!userForCheck) {
+          console.log(`Введён email ${email}, найден еmail ${userForCheck}`);
+          return { success: false, message: 'Wrong email' };
+        }
+
+        if (userForCheck.password === password) {
+          this.router.navigate(['/main']);
+          return { success: true, message: `Hello ${userForCheck.name}`, userId: userForCheck.id };
+        } else {
+          return { success: false, message: 'Password is wrong!' };
+        }
+      }),
+    );
   }
 
   getUser(id: number): Observable<any> {
